+++ src/Microsoft.Health.Fhir.SqlServer/Features/Schema/ScriptProvider.cs
@@ -1,0 +1,30 @@
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+using System.IO;
+using System.Reflection;
+
+namespace Microsoft.Health.Fhir.SqlServer.Features.Schema
+{
+    public static class ScriptProvider
+    {
+        public static string GetMigrationScript(int version)
+        {
+            string resourceName = $"{typeof(ScriptProvider).Namespace}.Migrations.{version}.sql";
+            using (Stream stream = Assembly.GetExecutingAssembly().GetManifestResourceStream(resourceName))
+            {
+                if (stream == null)
+                {
+                    throw new FileNotFoundException(Resources.ScriptNotFound);
+                }
+
+                using (var reader = new StreamReader(stream))
+                {
+                    return reader.ReadToEnd();
+                }
+            }
+        }
+    }
+}
